pyproject.toml
[tool.poetry]
name = "godot-asset-uploader"
version = "0.1.0"
description = "Automatically upload and update assets in Godot Asset Library"
authors = ["Maciej Katafiasz <mathrick@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
gdasset = "godot_asset_uploader.main:safe_cli"

[tool.poetry-pyinstaller-plugin]
use-poetry-install = true

[tool.poetry-pyinstaller-plugin.scripts]
gdasset = { source = "pyinstaller-main.py", type = "onefile", bundle = false}

[tool.poetry.dependencies]
# >=3.8.1 is because of flake8, <3.14 because of pyinstaller
python = "^3.8.1,<3.14"
mistletoe = "^1.4.0"
click = "^8.1.7"
dulwich = "^0.22.1"
requests = "^2.32.3"
validator-collection = "^1.5.0"
tomlkit = "^0.13.2"
giturlparse = "^0.12.0"
cloup = "^3.0.5"
dirtyjson = "^1.0.8"
yarl = "^1.13.1"
python-hglib = "^2.6.2"
decorators = "^2.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-regressions = "^2.5.0"
pyinstaller = "^6.10.0"
pytest-mock = "^3.14.0"
flake8 = "^7.1.1"
flake8-pyproject = "^1.2.3"

[tool.flake8]
max-line-length = 100
ignore = ["E401"]
exclude = [".git", "__pycache__"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
